Added Author Class and added complex one to one relation between the entities

+++ src/main/java/com/myspringboot/springbootfirst/controllers/BookController.java
@@ -1,8 +1,11 @@
 package com.myspringboot.springbootfirst.controllers;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
@@ -25,50 +28,93 @@ public class BookController {
     private BookService bookService;
 
     @RequestMapping(path = "/books", method = RequestMethod.GET)
-    public List<Book> getBooks() {
+    public ResponseEntity<List<Book>> getBooks() {
         // Static Custom Book creation and return
         // Book book = new Book();
         // book.setAuthor("John Doe");
         // book.setTitle("My Book");
         // book.setId(101);
 
-        List<Book> book = bookService.getAllBooks();
+        List<Book> books = bookService.getAllBooks();
 
-        return book;
+        if (books.size() <= 0) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
+        return ResponseEntity.of(Optional.of(books));
     }
 
     @GetMapping("/books/{id}")
-    public Book getBookById(@PathVariable("id") int id) {
+    public ResponseEntity<Book> getBookById(@PathVariable("id") int id) {
         Book book = bookService.getBookById(id);
-        return book;
+
+        if (book == null || book.getId() <= 0) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body(book);
     }
 
     @PutMapping("/books/{id}")
-    public Book updateBookById(@RequestBody Book book, @PathVariable("id") int id) {
+    public ResponseEntity<Book> updateBookById(@RequestBody Book book, @PathVariable("id") int id) {
 
-        Book result = bookService.updateBook(book, id);
-        return result;
+        Book result = null;
+        result = bookService.updateBook(book, id);
+        if (result != null)
+            return ResponseEntity.ok().body(result);
+        else
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 
     @PostMapping("/books")
-    public Book createBook(@RequestBody Book book) {
-        Book resultBook = bookService.saveBook(book);
-        return resultBook;
+    public ResponseEntity<Book> createBook(@RequestBody Book book) {
+        Book resultBook = null;
+        try {
+            resultBook = bookService.saveBook(book);
+            return ResponseEntity.status(HttpStatus.CREATED).body(resultBook);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+        }
     }
 
     @PostMapping("/nbooks")
-    public List<Book> createNBooks(@RequestBody List<Book> books) {
-        List<Book> resultBook = bookService.saveBooks(books);
-        return resultBook;
+    public ResponseEntity<List<Book>> createNBooks(@RequestBody List<Book> books) {
+        List<Book> resultBook = null;
+        try {
+            resultBook = bookService.saveBooks(books);
+            return ResponseEntity.status(HttpStatus.CREATED).body(resultBook);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+        }
     }
 
     @DeleteMapping("/books/{id}")
-    public void deleteBookById(@PathVariable("id") int id) {
-        bookService.deleteBook(id);
+    public ResponseEntity<String> deleteBookById(@PathVariable("id") int id) {
+        try {
+            if (bookService.getBookById(id) == null) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            }
+            bookService.deleteBook(id);
+            return ResponseEntity.status(HttpStatus.CREATED).body("Book Deleted!");
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+        }
     }
 
     @DeleteMapping("/books")
-    public void deleteAllBooks() {
-        bookService.deleteAllBooks();
+    public ResponseEntity<String> deleteAllBooks() {
+        try {
+            if (bookService.getAllBooks().size() <= 0) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            }
+            bookService.deleteAllBooks();
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("All Books Deleted!");
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+        }
     }
 }
